test(patina): split reserved component name tests
