+++ crates/vize_patina/src/linter/engine.rs
@@ -68,7 +68,6 @@ const SEMANTIC_TEMPLATE_RULES: &[&str] = &[
     "vue/no-mutating-props",
     "a11y/no-refer-to-non-existent-id",
 ];
-
 const SHARED_SFC_DESCRIPTOR_RULES: &[&str] = &[
     "vue/no-reserved-component-names",
     "vue/no-unused-refs",
